Adicionada tela de venda

+++ back-end/supermarket/src/main/java/com/supermarket/domain/dto/EstoqueDTO.java
@@ -16,5 +16,5 @@ public class EstoqueDTO {
 
     private String codigoBarras;
 
-    private Long qtdEstoque;
+    private Double qtdEstoque;
 }
